requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
deep-translator==1.11.4
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
filelock==3.20.3
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.1
groq==1.0.0
grpcio==1.76.0
gunicorn==25.1.0
h11==0.16.0
h2==3.2.0
h5py==3.15.1
hf-xet==1.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyboard==0.13.5
kiwisolver==1.4.9
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.11
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
namex==0.1.0
narwhals==2.16.0
networkx==3.4.2
nltk==3.9.2
numba==0.64.0
numpy==2.2.6
openai==2.23.0
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.2
propcache==0.4.1
protobuf==3.20.3
pyarrow==23.0.0
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyDirectInput==1.0.4
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.3.2
pypdf==6.7.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.3.1
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
secure-smtplib==0.1.1
sentence-transformers==5.2.2
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8.3
SQLAlchemy==2.0.46
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.3
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
watchdog==6.0.0
Werkzeug==3.1.5
wrapt==2.1.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
